+++ packages/f-form-field/src/components/FormField.vue
@@ -5,9 +5,11 @@
         <div
             :class="$style['c-formField-inputWrapper']">
             <form-label
-                v-if="normalisedLabelStyle === 'default'"
+                v-if="!isInline"
                 :label-style="normalisedLabelStyle"
-                :for="uniqueId">
+                :for="uniqueId"
+                :is-inline="isInline"
+                data-js-test="defaultLabel">
                 {{ labelText }}
             </form-label>
             <input
@@ -22,9 +24,11 @@
                 v-on="listeners"
             >
             <form-label
-                v-if="normalisedLabelStyle === 'inline'"
+                v-if="isInline"
                 :label-style="normalisedLabelStyle"
-                :for="uniqueId">
+                :for="uniqueId"
+                :is-inline="isInline"
+                data-js-test="inlineLabel">
                 {{ labelText }}
             </form-label>
         </div>
@@ -35,43 +39,64 @@
 <script>
 import { globalisationServices } from '@justeat/f-services';
 import FormLabel from './FormLabel.vue';
+import Debounce from '../services/debounce';
 import tenantConfigs from '../tenants';
-import { VALID_INPUT_TYPES, DEFAULT_INPUT_TYPE, VALID_LABEL_STYLES } from '../constants';
+import {
+    VALID_INPUT_TYPES,
+    DEFAULT_INPUT_TYPE,
+    VALID_LABEL_STYLES,
+    MOBILE_WIDTH
+} from '../constants';
 
 export default {
     name: 'FormField',
+
     components: {
         FormLabel
     },
+
     inheritAttrs: false,
+
     props: {
         locale: {
             type: String,
             default: ''
         },
+
         labelText: {
             type: String,
             default: ''
         },
+
         inputType: {
             type: String,
             default: DEFAULT_INPUT_TYPE,
             validator: value => (VALID_INPUT_TYPES.indexOf(value) !== -1) // The prop value must match one of the valid input types
         },
+
         labelStyle: {
             type: String,
             default: 'default',
             validator: value => (VALID_LABEL_STYLES.indexOf(value) !== -1) // The prop value must match one of the valid input types
         },
+
         value: {
             type: [String, Number],
             default: ''
         },
+
         dataTestId: {
             type: String,
             default: ''
         }
     },
+
+    data () {
+        return {
+            windowWidth: null
+        };
+    },
+
     computed: {
         normalisedInputType () {
             if (VALID_INPUT_TYPES.includes(this.inputType)) {
@@ -79,38 +104,66 @@ export default {
             }
             return DEFAULT_INPUT_TYPE;
         },
+
         normalisedLabelStyle () {
             if (VALID_LABEL_STYLES.includes(this.labelStyle)) {
                 return this.labelStyle;
             }
             return '';
         },
+
         formFieldLocale () {
             return globalisationServices.getLocale(tenantConfigs, this.locale, this.$i18n);
         },
+
         copy () {
             const localeConfig = tenantConfigs[this.formFieldLocale];
             return { ...localeConfig };
         },
+
         theme () {
             return globalisationServices.getTheme(this.formFieldLocale);
         },
+
         listeners () {
             return {
                 ...this.$listeners,
                 input: this.updateValue
             };
         },
+
         uniqueId () {
             return `formField-${(this.$attrs.name ? this.$attrs.name : this._uid)}`;
         },
+
         testId () {
             return this.dataTestId || this.$attrs.name || false;
+        },
+
+        isInline () {
+            return (this.windowWidth < MOBILE_WIDTH && this.labelStyle === 'inlineNarrow') ||
+                this.labelStyle === 'inline';
         }
     },
+
+    mounted () {
+        window.addEventListener('resize', Debounce(this.updateWidth, 100));
+        this.updateWidth();
+    },
+
+    destroyed () {
+        window.removeEventListener('resize', this.updateWidth);
+    },
+
     methods: {
         updateValue (event) {
             this.$emit('input', event.target.value);
+        },
+
+        updateWidth () {
+            if (typeof (window) !== 'undefined') {
+                this.windowWidth = window.innerWidth;
+            }
         }
     }
 };